+++ src/app/profile/bookmark/BookmarkList/BookmarkList.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/mixins';
+@import '../../../../scss/mixins';
 
 .container{
     display: flex;
