+++ src/pages/Profile/bookmark/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import BookmarkContainer from '@src/app/profile/bookmark/BookmarkContainer';
+
+const BookmarkPage = (): JSX.Element => {
+	return (
+		<div>
+			<BookmarkContainer />
+		</div>
+	);
+};
+
+export default BookmarkPage;
